+++ pkg/internal/ebpf/common/kafka_detect_transform.go
@@ -4,6 +4,8 @@ import (
 	"encoding/binary"
 	"errors"
 
+	trace2 "go.opentelemetry.io/otel/trace"
+
 	"github.com/grafana/beyla/pkg/internal/request"
 )
 
@@ -222,4 +224,38 @@ func getTopicOffsetFromFetchOperation(header *Header) int {
 	}
 
 	return offset
+}
+
+func TCPToKafkaToSpan(trace *TCPRequestInfo, data *KafkaInfo) request.Span {
+	peer := ""
+	hostname := ""
+	hostPort := 0
+
+	if trace.ConnInfo.S_port != 0 || trace.ConnInfo.D_port != 0 {
+		peer, hostname = trace.reqHostInfo()
+		hostPort = int(trace.ConnInfo.D_port)
+	}
+	return request.Span{
+		Type:           request.EventTypeKafkaClient,
+		Method:         data.Operation.String(),
+		OtherNamespace: data.ClientID,
+		Path:           data.Topic,
+		Peer:           peer,
+		Host:           hostname,
+		HostPort:       hostPort,
+		ContentLength:  0,
+		RequestStart:   int64(trace.StartMonotimeNs),
+		Start:          int64(trace.StartMonotimeNs),
+		End:            int64(trace.EndMonotimeNs),
+		Status:         0,
+		TraceID:        trace2.TraceID(trace.Tp.TraceId),
+		SpanID:         trace2.SpanID(trace.Tp.SpanId),
+		ParentSpanID:   trace2.SpanID(trace.Tp.ParentId),
+		Flags:          trace.Tp.Flags,
+		Pid: request.PidInfo{
+			HostPID:   trace.Pid.HostPid,
+			UserPID:   trace.Pid.UserPid,
+			Namespace: trace.Pid.Ns,
+		},
+	}
 }
